\name{aovQTL}
\alias{aovQTL}

\title{QTL mapping using ANOVA}

\usage{
   aovQTL(phenotype="", genotype="")
}

\arguments{
    \item{phenotype}{A data frame representing the phenotype of a population.}
    \item{genotype}{A data frame representing the genotype of a population.}
}

\description{
  Prints 'Hello, world!'.
}

\author{Wen Yao}

\examples{
 1+1
}
