\name{binQTLScan}
\alias{binQTLScan}

\title{QTL mapping with binmap data}

\usage{
   binQTLScan(phenotype="", genotype="", population="RIL")
}

\arguments{
    \item{phenotype}{A data frame of structural variations.}
    \item{genotype}{A data frame of structural variations.}
    \item{population}{A data frame of structural variations.}
}

\description{
Prints 'Hello, world!'.
}

\author{Wen Yao}

\examples{
 1 + 1
}
