\name{plotBinmap}
\alias{plotBinmap}

\title{Make a plot of Binmap}

\usage{
   plotBinmap(genotype="", cols=c("red", "blue", "grey40"), ...)
}

\arguments{
    \item{genotype}{A data frame of structural variations.}
    \item{cols}{Colours used to indicate different genotypes.}
    \item{...}{parameters passed to plot.}
}

\description{
Prints 'Hello, world!'.
}

\author{Wen Yao}

\examples{
 1 + 1
}
