\name{plotQTL}
\alias{plotQTL}

\title{Make a plot of QTL mapping result}

\usage{
   plotQTL(qtl.res="", y.bottom=0, y.top=2, cols=c("grey30", "grey70"),
           ylab="LOD", xlab="", ...)
}

\arguments{
    \item{qtl.res}{A data frame of structural variations.}
    \item{y.bottom}{A data frame of structural variations.}
    \item{y.top}{A data frame of structural variations.}
    \item{cols}{Colours used to indicate different chromosome.}
    \item{ylab}{Colours used to indicate different chromosome.}
    \item{xlab}{Colours used to indicate different chromosome.}
    \item{...}{parameters passed to plot.}
}

\description{
Prints 'Hello, world!'.
}

\author{Wen Yao}

\examples{
 1 + 1
}
